zsh export checks for fnm zoxide and pip for eval
Zsh sources exports script and enables some autocompletion capabilites.
For that it requires packages which might not be installed at the time
which results in error. To avoid that, I've added checks to see if these
commands are present and executable before running them and if they're
not, the shell should now ignore them.