Replace youtube-viewer with pipe-viewer for nboat
Newsboat is using youtube-viewer to search and view comments to videos
linked in the feed. Unfortunately it requires connection to youtube's
API in order to be able to view comments requiring user to go ahead and
create and register their app at google in order to request their API
key. This is hardly ideal as it introduces a dependency. Pipe-viewer is
a fork which connects to youtube directly instead.