Opus files should be moved to Music folder
the executable tidy is a bash script which scans the directory and then
executes mv command in order to move files to respective folders based
on the extension. This is especially useful to execute in Downloads
folder for example. When applicable I've started useing OPUS encoded
files which is supposedly successor to OGG Vorbis format offering
slightly better ballance in terms of quality/compression ratio.