Added mpv lua script to address no audio issue
When trying to play some videos from youtube (usually by opening them in
newsboat) sometimes the video would play with no audio. This issue is
related to Mpv's call to youtube-dl resulting in error. The issue was
addressed in latest build of mpv, however if you want to stay on stable
branch, you can fix the problem by using ytdl_hook.lua script instead.
In order for this script to work, you need to pass the --no-ytdl or
--ytdl=no flag to mpv. This flag will cause mpv to disable it's built-in
ytdl hook and look for alternative i.e. the ytdl_hook.lua script.