Modified the macros for newsboat.
The `<leader> + a` macro now uses yt-dlp's ability to use external
downloader. In this case it's `aria` - make sure to install the package
in order for this to work correctly.
The `<leader> + v` macro now adds the --no-ytdl flag for mpv to address
the issue where some youtube videos would play without audio.